requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
ftfy==6.3.1
grad-cam==1.5.4
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.1
pafy==0.5.5
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-gradcam==0.2.1
pytube==15.0.0
pytubefix==8.3.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torch-summary==1.4.5
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.2
ttach==0.0.3
typing_extensions==4.12.2
urllib3==2.2.3
vidgear==0.3.3
wcwidth==0.2.13
youtube-dl==2021.12.17
yt-dlp==2024.11.4
